multi-platform/bot/requirements.txt
python-telegram-bot==21.3
openai>=1.30.0
python-dotenv>=1.0.0
flask>=3.0.0
twilio>=9.0.0
gunicorn>=22.0.0
